Remove XML-RPC API
It's been deprecated since 2.0 with the new REST API. That API is
now pretty solid, and git-pw is good. Drop the old API.

Provide a page letting people know that the API is gone if they
access any of the old pages.

This breaks pwclient, which only supports the old API. So we delete
a few things that referred to it or used it, including some old tools.

Signed-off-by: Daniel Axtens <dja@axtens.net>